requirements.txt
lightrag-hku>=1.4.9.3   ##SMY: 1.3.6   --> 1.4.9.3
gradio[oauth]>=5.29.0           ##SMY: 4.0.0  (5.29.0)
networkx>=3.4.2         ##SMY: 3.2.1  (3.4.2)
pyvis>=0.3.2            ##SMY: 0.3.2
python-dotenv>=1.1.0    ##SMY: 1.1.0
openai>=2.3.0           ##SMY: 1.76.2  --> 2.3.0
ollama>=0.4.8           ##SMY: 0.4.8   --> 0.6.0
google-genai>=1.40.0    ##SMY: 1.43.0

plotly>=6.0.1           ##SMY: 6.0.1
tiktoken>=0.9.0         ##SMY: 0.9.0
graspologic>=3.4.1      ##SMY: 3.4.1 networkx: utilities and algorithms for processing and analysis of graphs

watchfiles>=1.0.5       ##SMY: 1.0.5
#globals                 ##SMY: